CR: Move local introspection to separate module.
